changes & new implementation
the server will not update the player body yaw, which means that on the client the bodyyaw gets displayed wrongly. Upon closing & re-entering the game, the client bodyYaw gets overwritten with the server bodyYaw. So we can conclude that the server bodyYaw is valid for other players. We need to take a server & client approach! Mixins have been disabled because they don't work on the server yet.