comments & modair implementation
still need to find a way to stop water from replacing it. Current API makes the method deprecated. The static class in Block: Block.Settings has methods that refer to that deprecated method. Since this is the only method to do it legally in the API, we must try to override that method to return false always.

Problem: the constructor is private...